requirements.txt
PySide6==6.5.2
matplotlib==3.7.2
pillow